Moved normalizeHtml to test utils.
